+++ src/rainsprite_state.c
@@ -1,5 +1,5 @@
 #include "../include/rainsprite_state.h"
-void rainsprite_state_init(RainspriteState * state, unsigned short sprite_num,unsigned int x_start, unsigned int y_start)
+void rainsprite_state_init(RainspriteState * state, unsigned short sprite_num,unsigned short x_start, unsigned short y_start)
 {
     state->is_exploding = 0x00;
     state->explosion_countdown = 180;
